🧹 Move pages to page folder
