+++ frontend/src/App.jsx
@@ -11,7 +11,7 @@ import { useTranslation } from 'react-i18next'
 import { CookiesProvider } from 'react-cookie'
 import GlobalTheme from './components/GlobalTheme'
 import TestPage from './components/TestPage'
-import HomePage from './HomePage'
+import HomePage from './pages/HomePage'
 import NotFoundPage from './NotFoundPage'
 import NotYetImplementedPage from './NotYetImplementedPage'
 import DialogProvider from './components/DialogProvider'
@@ -22,9 +22,9 @@ import VersionCheck from './components/VersionCheck'
 import AppFrame from './components/AppFrame'
 import BreakPointIndicator from './components/BreakPointIndicator'
 import SnackbarMessages from './components/SnackbarMessages'
-import ProfilePage from './ProfilePage'
-import InstallationsPage from './InstallationsPage'
-import DetailInstallationPage from './DetailInstallationPage'
+import ProfilePage from './pages/ProfilePage'
+import InstallationsPage from './pages/InstallationsPage'
+import DetailInstallationPage from './pages/DetailInstallationPage'
 
 const routes = [
   {
