+++ frontend/src/pages/InstallationsPage/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { useTranslation } from 'react-i18next'
-import dummyData from '../data/dummyinstallations.yaml'
+import dummyData from '../../data/dummyinstallations.yaml'
 import Button from '@mui/material/Button'
 import TableCell from '@mui/material/TableCell'
 import TableRow from '@mui/material/TableRow'
@@ -8,9 +8,9 @@ import Typography from '@mui/material/Typography'
 import MoreVertIcon from '@mui/icons-material/MoreVert'
 import SolarPowerIcon from '@mui/icons-material/SolarPower'
 import { Link } from 'react-router-dom'
-import TableEditor from '../components/TableEditor'
-import { useAuth } from '../components/AuthProvider'
-import ovapi from '../services/ovapi'
+import TableEditor from '../../components/TableEditor'
+import { useAuth } from '../../components/AuthProvider'
+import ovapi from '../../services/ovapi'
 import { Container } from '@mui/material'
 
 function PageTitle(props) {
